refactor: move files from `sidekick-new` to `sidekick`
