+++ src/apps/app-router.tsx
@@ -15,9 +15,9 @@ import { useSelector } from 'react-redux';
 import { RootState } from '../store/reducer';
 import { Provider as AuthenticationContextProvider } from '../components/authentication/context';
 
-import { Container as SidekickContainer } from '../components/sidekick-new/components/container';
-import { Header as SidekickHeader } from '../components/sidekick-new/components/header';
-import { CurrentUserDetails } from '../components/sidekick-new/components/current-user-details';
+import { Container as SidekickContainer } from '../components/sidekick/components/container';
+import { Header as SidekickHeader } from '../components/sidekick/components/header';
+import { CurrentUserDetails } from '../components/sidekick/components/current-user-details';
 
 const redirectToRoot = () => <Redirect to={'/'} />;
