+++ src/apps/feed/components/feed-chat/index.test.tsx
@@ -7,7 +7,7 @@ import { validateFeedChat } from '../../../../store/chat';
 import { send } from '../../../../store/messages';
 import { config } from '../../../../config';
 import { Spinner } from '@zero-tech/zui/components/LoadingIndicator';
-import { MembersSidekick } from '../../../../components/sidekick-new/variants/members-sidekick';
+import { MembersSidekick } from '../../../../components/sidekick/variants/members-sidekick';
 
 jest.mock('../../../../components/sidekick/variants/members-sidekick', () => ({
   MembersSidekick: () => <div data-testid='members-sidekick' />,
