--- src/components/sidekick-new/components/container/styles.module.scss
@@ -1,86 +1,0 @@
-@use '../../../../shared-components/theme-engine/theme' as theme;
-@use '../../../../modules/animation' as animation;
-
-@import '../../../../layout';
-@import '../../../../glass';
-
-.scroll-container__gradient {
-  background: linear-gradient(to bottom, transparent, theme.$background-color-tertiary-hover 100%);
-}
-
-.scroll-container__gradient-top {
-  background: linear-gradient(to top, transparent, theme.$background-color-tertiary-hover 100%);
-}
-
-.Container {
-  position: relative;
-  width: auto;
-  height: 100%;
-  transition: margin animation.$animation-duration-double var(--sidekick-transition-easing-function);
-  @include main-background;
-
-  box-sizing: border-box;
-  position: relative;
-  pointer-events: auto;
-  display: flex;
-  flex-direction: column;
-
-  flex-shrink: 0;
-  padding: 24px 18px 24px 0px;
-}
-
-.Wrapper {
-  box-sizing: border-box;
-  display: flex;
-  flex-shrink: 0;
-  pointer-events: auto;
-  min-width: $width-sidekick;
-  max-width: $width-sidekick;
-  width: 100%;
-  padding: 0;
-  flex: 1;
-}
-
-.Content {
-  display: flex;
-  flex-direction: column;
-
-  width: inherit;
-
-  border-radius: 16px;
-  overflow: hidden;
-
-  .Messages {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-  }
-
-  &[data-is-hidden] {
-    display: none;
-  }
-}
-
-.sidekick {
-  &--secondary {
-    overflow-y: auto;
-    max-width: 0;
-    overflow: hidden;
-
-    &.open {
-      transform: translateX(0);
-      max-width: 286px;
-      padding: 24px 0;
-      width: 100%;
-      margin-left: 18px;
-    }
-
-    &.close {
-      transform: translateX(100%);
-      max-width: 0;
-      padding: 24px 0;
-    }
-
-    transition: transform 0.1s ease-out, max-width 0.1s ease-out, padding 0.1s ease-out;
-  }
-}
