--- src/components/sidekick-new/components/current-user-details/components/user-header/styles.scss
@@ -1,73 +1,0 @@
-@use '~@zero-tech/zui/styles/theme' as theme;
-@import '../../../../../../variables';
-@import '../../../../../../glass';
-
-.user-header {
-  display: flex;
-  align-items: center;
-  height: 32px;
-  gap: 8px;
-  min-width: 0;
-
-  &__avatar-container {
-    cursor: pointer;
-  }
-
-  &__user-details {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    flex: 1 0 0;
-    width: 100%;
-    overflow: hidden;
-  }
-
-  &__name {
-    @include glass-text-primary-color;
-
-    font-size: $font-size-small;
-    font-weight: 600;
-    line-height: 15px;
-
-    width: 100%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-
-  &__handle {
-    @include glass-text-tertiary-color;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
-    gap: 4px;
-    max-width: 100%;
-
-    font-size: 10px;
-    line-height: 13px;
-
-    span {
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      max-width: 100%;
-    }
-  }
-
-  &__verify-id-button {
-    height: fit-content;
-    padding: 0;
-
-    * {
-      margin: 0;
-    }
-  }
-
-  &__verify-id-button-text {
-    font-size: 10px;
-    font-weight: 400;
-    line-height: 13px;
-  }
-}
