--- src/components/sidekick-new/components/header/styles.module.scss
@@ -1,12 +1,0 @@
-.Header {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  justify-content: space-between;
-  padding: 16px 16px 0 16px;
-}
-
-.Group {
-  display: flex;
-  gap: inherit;
-}
