--- src/components/sidekick-new/variants/members-sidekick.module.scss
@@ -1,46 +1,0 @@
-@use '../../../layout' as layout;
-@use '../../../variables' as variables;
-
-.Header {
-  max-width: layout.$width-sidekick;
-}
-
-.Members {
-  display: flex;
-  flex-direction: column;
-  max-width: 0;
-  overflow: hidden;
-  gap: variables.$default-gap;
-  height: 100%;
-  max-height: 100vh;
-
-  &.Open {
-    transform: translateX(0);
-    max-width: calc(layout.$width-sidekick + 16px);
-    padding: 24px 0;
-    width: 100%;
-  }
-
-  &.Closed {
-    transform: translateX(100%);
-    max-width: 0;
-    padding: 24px 0;
-  }
-
-  transition: transform 0.1s ease-out, max-width 0.1s ease-out, padding 0.1s ease-out;
-
-  > div {
-    flex-direction: column;
-  }
-}
-
-.Content {
-  flex: 1;
-  overflow-y: auto;
-  overflow-x: hidden;
-  height: 100%;
-  max-height: calc(100vh - 120px);
-  margin: 4px 4px 0 0;
-
-  scrollbar-width: none;
-}
