+++ src/components/sidekick/components/container/index.tsx
@@ -1,5 +1,8 @@
 import { IfAuthenticated } from '../../../authentication/if-authenticated';
 import { LegacyPanel } from '../../../layout/panel';
+import { SIDEKICK_PORTAL_ID } from '../../lib/constants';
+import { useSidekickContainer } from './useSidekickContainer';
+import { UserProfileContainer } from '../../../messenger/user-profile/container';
 
 import classNames from 'classnames';
 import styles from './styles.module.scss';
@@ -8,16 +11,23 @@ export interface ContainerProps {
   className?: string;
   children?: React.ReactNode;
   header?: React.ReactNode;
+  variant?: 'primary' | 'secondary';
 }
 
-export const Container = ({ className, children, header }: ContainerProps) => {
+export const Container = ({ className, children, header, variant = 'primary' }: ContainerProps) => {
+  const { isSettingsOpen } = useSidekickContainer(variant);
+
   return (
     <IfAuthenticated showChildren>
       <div className={classNames(styles.Container, className)}>
-        {header}
         <LegacyPanel className={styles.Wrapper}>
-          <div className={styles.Content}>
-            <div className={styles.Messages}>{children}</div>
+          {isSettingsOpen && <UserProfileContainer />}
+
+          {!isSettingsOpen && header}
+          {/* We're not removing the portal from the DOM here because it causes issues
+           * with the sidekick portal. */}
+          <div data-is-hidden={isSettingsOpen ? '' : null} className={styles.Content} id={SIDEKICK_PORTAL_ID}>
+            {children}
           </div>
         </LegacyPanel>
       </div>
