+++ src/components/sidekick/components/container/styles.module.scss
@@ -34,6 +34,7 @@
   display: flex;
   flex-shrink: 0;
   pointer-events: auto;
+  min-width: $width-sidekick;
   max-width: $width-sidekick;
   width: 100%;
   padding: 0;
@@ -41,6 +42,9 @@
 }
 
 .Content {
+  display: flex;
+  flex-direction: column;
+
   width: inherit;
 
   border-radius: 16px;
@@ -50,6 +54,10 @@
     height: 100%;
     display: flex;
     flex-direction: column;
+  }
+
+  &[data-is-hidden] {
+    display: none;
   }
 }
