+++ src/components/sidekick/components/content/index.tsx
