+++ src/components/sidekick/components/content/styles.module.scss
