+++ src/components/sidekick/components/scroll/index.tsx
