+++ src/components/sidekick/components/scroll/styles.module.scss
