+++ src/components/sidekick/lib/constants.ts
